final version 1

+++ AwawaTech.Mecanaut.API/DynamicMaintenancePlanning/Domain/Model/Aggregates/DynamicMaintenancePlan.cs
@@ -10,7 +10,7 @@ public class DynamicMaintenancePlan : AuditableAggregateRoot
     public string Name { get; private set; }
     public long MetricId { get; private set; }
     
-    public long Amount { get; private set; }
+    public double Amount { get; private set; }
     
     public long ProductionLineId { get; private set; }
     
@@ -20,7 +20,7 @@ public class DynamicMaintenancePlan : AuditableAggregateRoot
 
     protected DynamicMaintenancePlan() { }
 
-    private DynamicMaintenancePlan(string name, long metricId, long amount, long productionLineId, long plantLineId, TenantId tenantId)
+    private DynamicMaintenancePlan(string name, long metricId, double amount, long productionLineId, long plantLineId, TenantId tenantId)
     {
         Name = name;
         MetricId = metricId;
@@ -33,7 +33,7 @@ public class DynamicMaintenancePlan : AuditableAggregateRoot
         AddDomainEvent(new DynamicMaintenancePlanCreatedEvent(Id, tenantId.Value, name)); // Incluir Amount en el evento
     }
 
-    public static DynamicMaintenancePlan Create(string name, long metricId, long amount, long productionLineId, long plantLineId, TenantId tenantId)
+    public static DynamicMaintenancePlan Create(string name, long metricId, double amount, long productionLineId, long plantLineId, TenantId tenantId)
     {
         return new DynamicMaintenancePlan(name, metricId, amount, productionLineId, plantLineId, tenantId);
     }
